Move leaf_function tests into test_leaf_function.py (3/4)
[ghstack-poisoned]
