+++ test/dynamo/test_decorators.py
@@ -8,7 +8,6 @@ import unittest.mock as mock
 from unittest.mock import patch
 
 import torch
-import torch._dynamo.config as config
 import torch._dynamo.testing
 from torch._dynamo.decorators import leaf_function
 from torch._dynamo.exc import Unsupported
@@ -2448,397 +2447,6 @@ Detected recompile when torch.compile stance is 'fail_on_recompile'. filename: '
         model.forward = torch._dynamo.disable(model.forward, recursive=False)
         with self.assertRaises(RuntimeError):
             exported_model = torch.export.export(model, (inp,))
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    def test_leaf_function_input_mutation_error(self, backend):
-        @leaf_function
-        def mutate_input(x):
-            x.add_(1)
-            return (x,)
-
-        @mutate_input.register_fake
-        def mutate_input_fake(x):
-            x.add_(1)
-            return (x,)
-
-        def fn(x):
-            return mutate_input(x)
-
-        x = torch.randn(3, 3)
-
-        x_eager = x.clone()
-        with self.assertRaisesRegex(RuntimeError, "Undeclared in-place mutation"):
-            fn(x_eager)
-
-        x = torch.randn(3, 3, requires_grad=True)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        with self.assertRaisesRegex(RuntimeError, "leaf Variable that requires grad"):
-            compiled_fn(x.clone().requires_grad_(True))
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    def test_leaf_function_validation_dtype_mismatch(self, backend):
-        @leaf_function
-        def dtype_mismatch_forward(mod, x):
-            return (mod.linear(x),)
-
-        @dtype_mismatch_forward.register_fake
-        def dtype_mismatch_forward_fake(mod, x):
-            return (mod.linear(x).double(),)
-
-        class DtypeMismatchModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(3, 3)
-
-            def forward(self, x):
-                return dtype_mismatch_forward(self, x)
-
-        mod = DtypeMismatchModule()
-        x = torch.randn(3, 3)
-
-        with config.patch(leaf_function_validate_outputs=True):
-            compiled_mod = torch.compile(mod, backend=backend)
-            with self.assertRaisesRegex(RuntimeError, "Dtype mismatch"):
-                compiled_mod(x)
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    @parametrize("validate_outputs", [True, False])
-    def test_leaf_function_validation_shape_mismatch(self, backend, validate_outputs):
-        @leaf_function
-        def mismatched_forward(mod, x):
-            return (mod.linear(x),)
-
-        @mismatched_forward.register_fake
-        def mismatched_forward_fake(mod, x):
-            return (torch.zeros(x.shape[0], 6),)
-
-        class MismatchedModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(3, 3)
-
-            def forward(self, x):
-                return mismatched_forward(self, x)
-
-        mod = MismatchedModule()
-        x = torch.randn(3, 3)
-
-        with config.patch(leaf_function_validate_outputs=validate_outputs):
-            compiled_mod = torch.compile(mod, backend=backend)
-            if validate_outputs:
-                with self.assertRaises((RuntimeError, AssertionError)):
-                    compiled_mod(x)
-            else:
-                result = compiled_mod(x)
-                self.assertEqual(result[0].shape, (3, 3))
-
-    def test_leaf_function_no_module_inputs(self):
-        @leaf_function
-        def my_custom_fn(inputs: dict[str, torch.Tensor], scale: float, offset: int):
-            x = inputs["x"]
-            y = inputs["y"]
-            if x.sum() > 0:
-                return (x * scale + y + offset, x.sum() + y.sum())
-            return (x * scale - y + offset, x.sum() - y.sum())
-
-        @my_custom_fn.register_fake
-        def my_custom_fn_fake(
-            inputs: dict[str, torch.Tensor], scale: float, offset: int
-        ):
-            x = inputs["x"]
-            y = inputs["y"]
-            return (x * scale + y + offset, x.sum() + y.sum())
-
-        class NoModuleInputsModule(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.scale = 2.0
-                self.offset = 1
-
-            def forward(self, x, y):
-                inputs = {"x": x, "y": y}
-                return my_custom_fn(inputs, self.scale, self.offset)
-
-        def args_fn():
-            return (
-                torch.randn(3, 3, requires_grad=True),
-                torch.randn(3, 3, requires_grad=True),
-            )
-
-        def loss_fn(out):
-            return out[0].sum() + out[1].sum()
-
-        self._test_leaf_function_helper(NoModuleInputsModule, args_fn, loss_fn)
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    @parametrize("check_escaped_gradients", [True, False])
-    def test_leaf_function_escaped_gradient_multiple_tensors(
-        self, backend, check_escaped_gradients
-    ):
-        weight1 = torch.randn(3, 3, requires_grad=True)
-        weight2 = torch.randn(3, 3, requires_grad=True)
-
-        @leaf_function
-        def uses_multiple_closures(x):
-            return (x @ weight1 + x @ weight2,)
-
-        @uses_multiple_closures.register_fake
-        def uses_multiple_closures_fake(x):
-            return (torch.empty(x.shape[0], 3),)
-
-        def fn(x):
-            return uses_multiple_closures(x)
-
-        x = torch.randn(2, 3, requires_grad=True)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        with config.patch(
-            leaf_function_check_escaped_gradients=check_escaped_gradients
-        ):
-            if check_escaped_gradients:
-                with self.assertRaisesRegex(RuntimeError, "2 tensor"):
-                    compiled_fn(x)
-            else:
-                result = compiled_fn(x)
-                self.assertEqual(result[0].shape, (2, 3))
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    @parametrize("check_escaped_gradients", [True, False])
-    def test_leaf_function_escaped_gradient_input_no_grad(
-        self, backend, check_escaped_gradients
-    ):
-        closure_weight = torch.randn(3, 3, requires_grad=True)
-
-        @leaf_function
-        def uses_closure(x):
-            return (x @ closure_weight,)
-
-        @uses_closure.register_fake
-        def uses_closure_fake(x):
-            return (torch.empty(x.shape[0], 3),)
-
-        def fn(x):
-            return uses_closure(x)
-
-        x = torch.randn(2, 3, requires_grad=False)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        with config.patch(
-            leaf_function_check_escaped_gradients=check_escaped_gradients
-        ):
-            result = compiled_fn(x)
-            self.assertEqual(result[0].shape, (2, 3))
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    @parametrize("check_escaped_gradients", [True, False])
-    def test_leaf_function_escaped_gradient_mixed_inputs(
-        self, backend, check_escaped_gradients
-    ):
-        base1 = torch.randn(3, 3, requires_grad=True)
-        base2 = torch.randn(3, 4, requires_grad=True)
-        closure_weight1 = base1 * 2
-        closure_weight2 = base2 * 3
-
-        @leaf_function
-        def mixed_inputs(x, y):
-            out1 = x @ closure_weight1 + y
-            out2 = x @ closure_weight2
-            return (out1, out2)
-
-        @mixed_inputs.register_fake
-        def mixed_inputs_fake(x, y):
-            return (torch.empty(x.shape[0], 3), torch.empty(x.shape[0], 4))
-
-        def fn(x, y):
-            return mixed_inputs(x, y)
-
-        x = torch.randn(2, 3, requires_grad=True)
-        y = torch.randn(2, 3, requires_grad=False)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        with config.patch(
-            leaf_function_check_escaped_gradients=check_escaped_gradients
-        ):
-            if check_escaped_gradients:
-                with self.assertRaisesRegex(RuntimeError, "2 tensor"):
-                    compiled_fn(x, y)
-            else:
-                result = compiled_fn(x, y)
-                self.assertEqual(result[0].shape, (2, 3))
-                self.assertEqual(result[1].shape, (2, 4))
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    def test_leaf_function_escaped_gradient_error_message_contains_tensor_info(
-        self, backend
-    ):
-        closure_weight = torch.randn(4, 5, dtype=torch.float32, requires_grad=True)
-
-        @leaf_function
-        def uses_closure(x):
-            return (x @ closure_weight,)
-
-        @uses_closure.register_fake
-        def uses_closure_fake(x):
-            return (torch.empty(x.shape[0], 5),)
-
-        def fn(x):
-            return uses_closure(x)
-
-        x = torch.randn(2, 4, requires_grad=True)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        with config.patch(leaf_function_check_escaped_gradients=True):
-            with self.assertRaisesRegex(RuntimeError, r"shape=\[4, 5\].*dtype="):
-                compiled_fn(x)
-
-    @parametrize("backend", ["eager", "aot_eager"])
-    def test_leaf_function_escaped_gradient_actually_lost(self, backend):
-        closure_weight = torch.randn(3, 3, requires_grad=True)
-
-        @leaf_function
-        def uses_closure(x):
-            return (x @ closure_weight,)
-
-        @uses_closure.register_fake
-        def uses_closure_fake(x):
-            return (torch.empty(x.shape[0], 3),)
-
-        def fn(x):
-            return uses_closure(x)
-
-        x = torch.randn(2, 3, requires_grad=True)
-
-        compiled_fn = torch.compile(fn, backend=backend, fullgraph=True)
-        result = compiled_fn(x)
-        loss = result[0].sum()
-        loss.backward()
-
-        self.assertIsNotNone(x.grad)
-        self.assertIsNone(closure_weight.grad)
-
-    def test_leaf_function_and_nonstrict_trace_mutually_exclusive(self):
-        from torch._dynamo.decorators import leaf_function, nonstrict_trace
-
-        with self.assertRaisesRegex(
-            ValueError,
-            "cannot be both marked as @leaf_function and @nonstrict_trace",
-        ):
-
-            @leaf_function
-            @nonstrict_trace
-            def bad_fn1(x):
-                return (x,)
-
-        with self.assertRaisesRegex(
-            ValueError,
-            "cannot be both marked as @leaf_function and @nonstrict_trace",
-        ):
-
-            @nonstrict_trace
-            @leaf_function
-            def bad_fn2(x):
-                return (x,)
-
-    def test_leaf_function_no_return_value(self):
-        printed = []
-
-        @leaf_function
-        def fn_no_return(x):
-            print("processing")
-
-        @fn_no_return.register_fake
-        def fn_no_return_fake(x):
-            pass
-
-        class Mod(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = torch.nn.Linear(3, 3)
-
-            def forward(self, x):
-                fn_no_return(x)
-                return (self.linear(x),)
-
-        def args_fn():
-            return (torch.randn(3, 3, requires_grad=True),)
-
-        def loss_fn(out):
-            return out[0].sum()
-
-        with patch("builtins.print", lambda *args, **kwargs: printed.append(args)):
-            eager_graph, fw_graph, bw_graph = self._test_leaf_function_helper(
-                Mod, args_fn, loss_fn
-            )
-        self.assertTrue(any("processing" in p for p in printed))
-        self.assertExpectedInline(
-            eager_graph,
-            """\
-class GraphModule(torch.nn.Module):
-    def forward(self, L_x_: "f32[3, 3]", L_self_modules_linear_parameters_weight_: "f32[3, 3]", L_self_modules_linear_parameters_bias_: "f32[3]"):
-        l_x_ = L_x_
-        l_self_modules_linear_parameters_weight_ = L_self_modules_linear_parameters_weight_
-        l_self_modules_linear_parameters_bias_ = L_self_modules_linear_parameters_bias_
-
-        real_fn : torch._higher_order_ops.invoke_leaf_function._LeafCallable = self.real_fn
-        fake_fn : torch._higher_order_ops.invoke_leaf_function._LeafCallable = self.fake_fn
-        input_spec : torch.utils._pytree.TreeSpec = self.input_spec
-        invoke_leaf_function = torch.ops.higher_order.invoke_leaf_function(real_fn, fake_fn, input_spec, '', l_x_);  real_fn = fake_fn = input_spec = invoke_leaf_function = None
-
-        linear: "f32[3, 3]" = torch._C._nn.linear(l_x_, l_self_modules_linear_parameters_weight_, l_self_modules_linear_parameters_bias_);  l_x_ = l_self_modules_linear_parameters_weight_ = l_self_modules_linear_parameters_bias_ = None
-        return (linear,)
-""",  # noqa: B950
-        )
-        self.assertExpectedInline(
-            fw_graph,
-            """\
-class GraphModule(torch.nn.Module):
-    def forward(self, primals_1: "f32[0]", primals_2: "f32[3, 3]", primals_3: "f32[3, 3]", primals_4: "f32[3]"):
-        _opaque_obj0 = self._opaque_obj0
-        _opaque_obj1 = self._opaque_obj1
-        _tree_spec_constant0 = self._tree_spec_constant0
-        with_effects = torch.ops.higher_order.with_effects(primals_1, torch.ops.higher_order.invoke_leaf_function, _opaque_obj0, _opaque_obj1, _tree_spec_constant0, '', primals_2, requires_grad_indices = (0,));  primals_1 = _opaque_obj0 = _opaque_obj1 = _tree_spec_constant0 = None
-
-        getitem: "f32[0]" = with_effects[0];  with_effects = None
-
-        t: "f32[3, 3]" = torch.ops.aten.t.default(primals_3)
-        addmm: "f32[3, 3]" = torch.ops.aten.addmm.default(primals_4, primals_2, t);  primals_4 = t = None
-        return (getitem, addmm, primals_2, primals_3)
-""",  # noqa: B950
-        )
-        self.assertExpectedInline(
-            bw_graph,
-            """\
-class GraphModule(torch.nn.Module):
-    def forward(self, primals_2: "f32[3, 3]", primals_3: "f32[3, 3]", tangents_1: "f32[3, 3]"):
-        t: "f32[3, 3]" = torch.ops.aten.t.default(primals_3);  primals_3 = None
-        t_1: "f32[3, 3]" = torch.ops.aten.t.default(t);  t = None
-        mm: "f32[3, 3]" = torch.ops.aten.mm.default(tangents_1, t_1);  t_1 = None
-        t_2: "f32[3, 3]" = torch.ops.aten.t.default(tangents_1)
-        mm_1: "f32[3, 3]" = torch.ops.aten.mm.default(t_2, primals_2);  t_2 = primals_2 = None
-        t_3: "f32[3, 3]" = torch.ops.aten.t.default(mm_1);  mm_1 = None
-        sum_1: "f32[1, 3]" = torch.ops.aten.sum.dim_IntList(tangents_1, [0], True);  tangents_1 = None
-        view: "f32[3]" = torch.ops.aten.view.default(sum_1, [3]);  sum_1 = None
-        t_4: "f32[3, 3]" = torch.ops.aten.t.default(t_3);  t_3 = None
-        return (mm, t_4, view)
-""",  # noqa: B950
-        )
-
-    def test_leaf_function_output_structure_mismatch(self):
-        @leaf_function
-        def mismatched_fn(x):
-            return {"a": x, "b": x * 2}
-
-        @mismatched_fn.register_fake
-        def mismatched_fn_fake(x):
-            return (x, x * 2)
-
-        def fn(x):
-            return mismatched_fn(x)
-
-        x = torch.randn(3, 3)
-        with self.assertRaisesRegex(AssertionError, "output structure mismatch"):
-            torch.compile(fn, backend="eager")(x)
 
     def test_leaf_function_nested_output(self):
         @leaf_function
